fix: computednavlinks logic moved to component
