+++ src/service/usersService.js
@@ -1,6 +1,6 @@
 import axios from 'axios'
 const apiClient = axios.create({
-  baseURL: 'http://127.0.0.1:8080',
+  baseURL: 'http://localhost:8080',
   headers: {
     Accept: 'application/json',
     'Content-Type': 'application/json'
